refactor(pytest): refactor test fixtures into separate files
This commit reorganizes the test fixtures by moving them from a single
conftest.py file into individual files within a new 'fixtures'
directory. Each fixture file is now dedicated to a specific component
or functionality, improving modularity and maintainability of the test
suite.

Key changes:
- Create new 'fixtures' directory under 'tests'
- Split fixtures into separate files based on functionality
- Update conftest.py to use pytest_plugins for fixture discovery
- Remove duplicate fixture definitions
- Improve import organization and remove unused imports
