+++ tests/document_test.py
@@ -4,7 +4,7 @@ from __future__ import annotations
 
 import requests_mock
 
-from tests.conftest import Company
+from tests.fixtures.document_fixtures import Companies
 from tests.utils.object_assertions import (
     assert_match_object,
     assert_object_lists_match,
@@ -34,7 +34,7 @@ def test_init(fake_api_call: ApiCall) -> None:
 
 def test_retrieve(fake_document: Document) -> None:
     """Test that the Document object can retrieve an document."""
-    json_response: Company = {
+    json_response: Companies = {
         "company_name": "Company",
         "id": "0",
         "num_employees": 10,
@@ -59,7 +59,7 @@ def test_retrieve(fake_document: Document) -> None:
 
 def test_delete(fake_document: Document) -> None:
     """Test that the Document object can delete an document."""
-    json_response: Company = {
+    json_response: Companies = {
         "company_name": "Company",
         "id": "0",
         "num_employees": 10,
